requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.7.0
click==8.1.8
fastapi==0.115.6
h11==0.14.0
idna==3.10
imageio==2.36.1
lazy-loader==0.4
networkx==3.4.2
numpy==2.2.1
opencv-python-headless==4.10.0.84
packaging==24.2
pillow==11.0.0
pydantic==2.10.4
pydantic-core==2.27.2
python-multipart==0.0.20
scikit-image==0.25.0
scipy==1.14.1
sniffio==1.3.1
starlette==0.41.3
tifffile==2024.12.12
typing-extensions==4.12.2
uvicorn==0.34.0
